refactor: use default template to add custom sections to foxy-customer in foxy-customer-portal
